requirements.txt
openai
numpy
Pillow
pandas
tqdm
python-dotenv
matplotlib
seaborn